fix: resolve CI test scope and web type-check errors
Scope flutter test to packages that have tests (run_recorder,
mobile_android). Switch web lint step to svelte-check. Add explicit
type annotations in data.ts and supabase-server.ts to satisfy strict
type checking.

+++ apps/web/src/lib/data.ts
@@ -67,7 +67,7 @@ async function decompressGzip(buf: ArrayBuffer): Promise<Uint8Array> {
 	while (true) {
 		const { done, value } = await reader.read();
 		if (done) break;
-		chunks.push(value);
+		chunks.push(value as Uint8Array);
 	}
 	const total = chunks.reduce((a, c) => a + c.length, 0);
 	const out = new Uint8Array(total);
